+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 [GD-2506](https://jira.gutefrage.net/browse/GD-2506) Ensure that display is only called once.
 
+## 3.12.1
+
+[GD-2447](https://jira.gutefrage.net/browse/GD-2447) Add possibility to configure Zeotap script params with an include
+list for key/values, and to prevent data collection using an exclude list.
+
 ## 3.12.0
 
 [GD-2447](https://jira.gutefrage.net/browse/GD-2447) Zeotap module added, supporting their data collection and id